new demo view
